+++ application/configs/environment.php
@@ -16,4 +16,4 @@ $environment = 'production';
 // Version number for current release
 $version_no = 'v1.14';
 // Release date for current release
-$version_release_date = 'In Progress';
+$version_release_date = '23rd February 2017';