+++ library/Dlayer/View/FormElementQuill.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * View helper for Quill rich text field
+ *
+ * @author Dean Blackborough <dean@g3d-development.com>
+ * @copyright G3D Development Limited
+ * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
+ */
+class Dlayer_View_FormElementQuill extends Zend_View_Helper_FormElement
+{
+    public function formElementQuill($name, $value = null, $attribs = null)
+    {
+        $info = $this->_getInfo($name, $value, $attribs);
+        extract($info); // name, value, attribs, options, listsep, disable
+
+        // build the element
+        $disabled = '';
+        if ($disable) {
+            // disabled
+            $disabled = ' disabled="disabled"';
+        }
+
+        $xhtml = '<input type="hidden"'
+            . ' name="' . $this->view->escape($name) . '"'
+            . ' id="' . $this->view->escape($id) . '"'
+            . ' value="' . $this->view->escape($value) . '"'
+            . $disabled
+            . $this->_htmlAttribs($attribs)
+            . $this->getClosingBracket()
+            . '<div id="editor-container-' . $this->view->escape($id) . '" class="ql-editor-container"></div>';
+
+        return $xhtml;
+    }
+}